Move encode/decode to a separate file
