+++ messenger/codec.go
@@ -1,0 +1,17 @@
+package messenger
+
+import (
+	"bytes"
+	"github.com/ugorji/go/codec"
+)
+
+var ch codec.CborHandle
+
+func encode(v interface{}, buf *bytes.Buffer) {
+	codec.NewEncoder(buf, &ch).MustEncode(v)
+}
+
+func decode(buf *bytes.Buffer, v interface{}) {
+	dec := codec.NewDecoder(buf, &ch)
+	dec.MustDecode(v)
+}
